Use BigIntegerField instead of PositiveIntegerField

+++ sequences/migrations/0001_initial.py
@@ -17,7 +17,7 @@ class Migration(migrations.Migration):
             name='Sequence',
             fields=[
                 ('name', models.CharField(max_length=100, primary_key=True, serialize=False, verbose_name='name')),
-                ('last', models.PositiveIntegerField(verbose_name='last value')),
+                ('last', models.BigIntegerField(verbose_name='last value')),
             ],
             options={
                 'verbose_name': 'sequence',
